added a functioning modal to profile page
